Handle keyboard interrupt when executing subprocesses
KeyboardInterrupt does not inherit from Exception, so it needs to be
handled separately.  If the controlling Python script catches it, it
will need to ensure the subprocess is killed as well, so mechanisms to
ensure that have been added.

Test-Information:

Interrupting a build of Swiften is now noted, build stopped.